requirements.txt
django
PIL
south
IPython
psycopg2
django-haystack
pysolr
